admin

+++ admin/views/producttype.php
@@ -6,45 +6,50 @@ $path_producttypemodel_delete   = "models/producttypemodel/typedelete.php";
 ?>
 
 <!--Upload Detail Product-->
-<form action="<?php echo $path_producttypemodel_input; ?>" id="typeinput" enctype="multipart/form-data" name="typeinput" method="post">
-    <table border="0" cellpadding="3" cellspacing="1" bgcolor="#CCCCCC">
-        <tr bgcolor="#FFFFFF">
-            <td>Name type</td>
-            <td><input type="text" name="product_type_name" id="product_type_name" value="" required></td>
-        </tr>
-        <tr align="right" bgcolor="#FFFFFF">
-            <td colspan="2"><input type="submit" name="typesubmit" value="Create"></td>
-        </tr>
-    </table>
-</form>
+<div class="col-xs-12 col-sm-6 col-md-6">
+    <div class="thumbnail">
+        <h3>Name type</h3>
+        <form class="form-horizontal" action="<?php echo $path_producttypemodel_input; ?>" id="typeinput" enctype="multipart/form-data" name="typeinput" method="post">
+            <div class="form-group">
+                <div class="col-xs-10">
+                    <input class="form-control" name="product_type_name" id="product_type_name" value="" required>
+                </div>
+                <div class="col-xs-1">
+                    <button class="btn btn-default" type="submit" name="typesubmit">
+                        <i class="far fa-save"></i>
+                    </button>
+                </div>
+            </div>
+        </form>
+    </div>
+    <br>
 
-<br>
+    <div class="thumbnail" style="height:270px; width:100%; overflow: auto;">
 
-<?php
-    include $path_basemodel;
-    $sql = "SELECT * FROM tb_product_type";
-    $result = mysqli_query($connection, $sql);
-?>
-<div style=" width:250px; height:100px; overflow: auto;">
-    <form>
-        <table border="0" cellpadding="3" cellspacing="1" bgcolor="#CCCCCC">
-            <tr align="center" bgcolor="#FFFFFF">
-                <td>Type </td>
-                <td>Delete</td>
-            </tr>
-            <?php
-            while ($row = mysqli_fetch_array($result)) {
-            ?>
+        <?php
+        include $path_basemodel;
+        $sql = "SELECT * FROM tb_product_type";
+        $result = mysqli_query($connection, $sql);
+        ?>
+        <form>
+            <table border="0" cellpadding="3" cellspacing="1" bgcolor="#CCCCCC">
                 <tr align="center" bgcolor="#FFFFFF">
-                    <td hidden><?php echo $row["product_type_id"] ?></td>
-                    <td><?php echo $row["product_type_name"]; ?></td>
-                    <td><a href="models/producttypemodel/typedelete.php?delete_id=<?php echo $row["product_type_id"]; ?>" 
-                    onclick="return confirm('โปรดตรวจสอบ Sub-type ก่อนลบ')">ลบ</a></td>
+                    <td>Type </td>
+                    <td>Delete</td>
                 </tr>
-            <?php
-            }
-            mysqli_close($connection);
-            ?>
-        </table>
-    </form>
+                <?php
+                while ($row = mysqli_fetch_array($result)) {
+                ?>
+                    <tr align="center" bgcolor="#FFFFFF">
+                        <td hidden><?php echo $row["product_type_id"] ?></td>
+                        <td><?php echo $row["product_type_name"]; ?></td>
+                        <td><a href="models/producttypemodel/typedelete.php?delete_id=<?php echo $row["product_type_id"]; ?>" onclick="return confirm('โปรดตรวจสอบ Sub-type ก่อนลบ')">ลบ</a></td>
+                    </tr>
+                <?php
+                }
+                mysqli_close($connection);
+                ?>
+            </table>
+        </form>
+    </div>
 </div>
